+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * feat: 依赖库升级
 * feat: WrapperCachedNetworkImage新增defaultPlaceholder属性
+* feat: Demo升级支持AGP8.0
+* feat: TabBar indicator 支持背景渐变和背景图片
 
 ## 1.2.0